Use strtol more, and add checks when atoi is used

+++ dtls.h
@@ -32,9 +32,9 @@ const char *janus_get_ssl_version(void);
  * @param[in] server_pem Path to the certificate to use
  * @param[in] server_key Path to the key to use
  * @param[in] password Password needed to use the key, if any
- * @param[in] timeout DTLS timeout base to use for retransmissions (ignored if not using BoringSSL)
+ * @param[in] timeout DTLS timeout base, in ms, to use for retransmissions (ignored if not using BoringSSL)
  * @returns 0 in case of success, a negative integer on errors */
-gint janus_dtls_srtp_init(const char *server_pem, const char *server_key, const char *password, guint timeout);
+gint janus_dtls_srtp_init(const char *server_pem, const char *server_key, const char *password, guint16 timeout);
 /*! \brief Method to cleanup DTLS stuff before exiting */
 void janus_dtls_srtp_cleanup(void);
 /*! \brief Method to return a string representation (SHA-256) of the certificate fingerprint */
